+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Update `RURL#href` everytime `RURL#addQuery` is called
 * Add `Date.parseDuration(str)` for getting a duration in ms
+* `Date#add` and `Date#subtract` now also accept duration strings
 
 ## 0.5.6 (2018-06-20)